// Rio.Web/src/app/shared/generated/model/trade-activity-by-month-dto.ts
/**
 * Rio.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class TradeActivityByMonthDto { 
    GroupingDate?: string;
    TradeVolume?: number;
    MinimumPrice?: number;
    MaximumPrice?: number;
    AveragePrice?: number;
    NumberOfTrades?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}